+++ components/core/preference/src/main/java/com/flipperdevices/core/preference/PreferenceDefault.kt
@@ -1,0 +1,5 @@
+package com.flipperdevices.core.preference
+
+object PreferenceDefault {
+    const val DEFAULT_UPDATE_URL = "https://update.flipperzero.one/firmware/directory.json"
+}