+++ components/settings/impl/src/main/java/com/flipperdevices/settings/impl/viewmodels/ExperimentalViewModel.kt
@@ -1,19 +1,24 @@
 package com.flipperdevices.settings.impl.viewmodels
 
+import androidx.datastore.core.DataStore
+import androidx.lifecycle.viewModelScope
 import androidx.navigation.NavController
 import com.flipperdevices.core.di.provideDelegate
+import com.flipperdevices.core.preference.pb.Settings
 import com.flipperdevices.core.ui.lifecycle.LifecycleViewModel
 import com.flipperdevices.filemanager.api.navigation.FileManagerEntry
 import com.flipperdevices.metric.api.MetricApi
 import com.flipperdevices.metric.api.events.SimpleEvent
 import com.flipperdevices.screenstreaming.api.ScreenStreamingFeatureEntry
 import javax.inject.Provider
+import kotlinx.coroutines.launch
 import tangle.viewmodel.VMInject
 
 class ExperimentalViewModel @VMInject constructor(
     fileManagerEntryProvider: Provider<FileManagerEntry>,
     metricApiProvider: Provider<MetricApi>,
-    private val screenStreamingFeatureEntry: ScreenStreamingFeatureEntry
+    private val screenStreamingFeatureEntry: ScreenStreamingFeatureEntry,
+    private val dataStore: DataStore<Settings>
 ) : LifecycleViewModel() {
     private val fileManagerEntry by fileManagerEntryProvider
     private val metricApi by metricApiProvider
@@ -26,5 +31,15 @@ class ExperimentalViewModel @VMInject constructor(
     fun onOpenScreenStreaming(navController: NavController) {
         metricApi.reportSimpleEvent(SimpleEvent.EXPERIMENTAL_OPEN_SCREENSTREAMING)
         navController.navigate(screenStreamingFeatureEntry.ROUTE.name)
+    }
+
+    fun onEditUpdateUrl(newUrl: String) {
+        viewModelScope.launch {
+            dataStore.updateData {
+                it.toBuilder()
+                    .setCustomUpdateUrl(newUrl)
+                    .build()
+            }
+        }
     }
 }