+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - [Feature] Rework all system dialog to custom
 - [Feature] New share flow(with `sf#path=()&id=()key=`)
 - [Feature] Share shadow file if that exist on NFC
+- [Feature] Add firmware update custom url
 - [FIX] Uploading share errors
 - [FIX] Display by white color meta data on NFC card